+++ ProjetoFinal/src/java/br/com/projeto/bean/VendasBean.java
@@ -48,7 +48,7 @@ public class VendasBean {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
     
-    public float valorTotal(){
+    public double valorTotal(){
         return quantidade*produto.getPreco();
     }
